Test pass with Observable.just()
Why I added that branch. Previously, me and Emre, tried to figure out the reason why test pass successfully but the app doesn't work.
The first attempt to reproduce the case when merging
RxRelay with RxBinding, was to merge RxRelay with Observable.just() in the test, but it didn't work.

This approach didn't work, because Observable.just() will triggers for every individual observer.